requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.8.3
aiosignal==1.4.0
altgraph==0.17.4
anyio==4.1.0
arrow==1.3.0
async-timeout==4.0.3
attrs==23.1.0
binaryornot==0.4.4
blinker==1.7.0
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
cookiecutter==2.5.0
distlib==0.3.8
distro==1.8.0
filelock==3.13.1
Flask==3.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
flet==0.14.0
flet-core==0.14.0
flet-runtime==0.14.0
frozenlist==1.7.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jpype1==1.6.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.6.3
numpy==1.26.1
oauthlib==3.2.2
packaging==23.2
pandas==2.1.2
pefile==2023.2.7
platformdirs==4.1.0
propcache==0.3.2
pygame==2.5.2
Pygments==2.17.2
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
PyJWT==2.8.0
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.1.1
python-slugify==8.0.1
pytz==2023.3.post1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
qrcode==7.4.2
repath==0.9.0
requests==2.31.0
rich==13.7.0
setuptools==80.9.0
six==1.16.0
sniffio==1.3.0
tabula-py==2.8.2
text-unidecode==1.3
twilio==8.10.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
virtualenv==20.25.0
watchdog==3.0.0
websocket-client==1.6.4
websockets==11.0.3
Werkzeug==3.1.3
yarl==1.20.1
